requirements.txt
google-genai
python-dotenv
streamlit
pandas
openpyxl
gTTS
pytest